# eds: simplify and optimize mTLS checker