--- a/pilot/pkg/networking/util/util.go
+++ b/pilot/pkg/networking/util/util.go
@@ -34,7 +34,6 @@ import (
 	httpv3 "github.com/envoyproxy/go-control-plane/envoy/type/http/v3"
 	matcher "github.com/envoyproxy/go-control-plane/envoy/type/matcher/v3"
 	"github.com/envoyproxy/go-control-plane/pkg/wellknown"
-	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/anypb"
 	"google.golang.org/protobuf/types/known/durationpb"
 	"google.golang.org/protobuf/types/known/structpb"
@@ -503,53 +502,6 @@ func AppendLbEndpointMetadata(istioMetadata *model.EndpointMetadata, envoyMetada
 		sb.WriteString(istioMetadata.ClusterID.String())
 		addIstioEndpointLabel(envoyMetadata, "workload", &structpb.Value{Kind: &structpb.Value_StringValue{StringValue: sb.String()}})
 	}
-}
-
-// MaybeApplyTLSModeLabel may or may not update the metadata for the Envoy transport socket matches for auto mTLS.
-func MaybeApplyTLSModeLabel(ep *endpoint.LbEndpoint, tlsMode string) (*endpoint.LbEndpoint, bool) {
-	if ep == nil || ep.Metadata == nil {
-		return nil, false
-	}
-	epTLSMode := ""
-	if ep.Metadata.FilterMetadata != nil {
-		if _, f := ep.Metadata.FilterMetadata[EnvoyTransportSocketMetadataKey].GetFields()[model.TunnelLabelShortName]; f {
-			// Tunnel > MTLS
-			return nil, false
-		}
-		if v, ok := ep.Metadata.FilterMetadata[EnvoyTransportSocketMetadataKey]; ok {
-			epTLSMode = v.Fields[model.TLSModeLabelShortname].GetStringValue()
-		}
-	}
-	// Normalize the tls label name before comparison. This ensure we won't falsely cloning
-	// the endpoint when they are "" and model.DisabledTLSModeLabel.
-	if epTLSMode == model.DisabledTLSModeLabel {
-		epTLSMode = ""
-	}
-	if tlsMode == model.DisabledTLSModeLabel {
-		tlsMode = ""
-	}
-	if epTLSMode == tlsMode {
-		return nil, false
-	}
-	// We make a copy instead of modifying on existing endpoint pointer directly to avoid data race.
-	// See https://github.com/istio/istio/issues/34227 for details.
-	newEndpoint := proto.Clone(ep).(*endpoint.LbEndpoint)
-
-	// ep.Metadata.FilterMetadata maybe an empty map or nil, after clone, corresponding field will be a nil map.
-	if newEndpoint.Metadata.FilterMetadata == nil {
-		newEndpoint.Metadata.FilterMetadata = make(map[string]*structpb.Struct)
-	}
-
-	if tlsMode != "" && tlsMode != model.DisabledTLSModeLabel {
-		newEndpoint.Metadata.FilterMetadata[EnvoyTransportSocketMetadataKey] = &structpb.Struct{
-			Fields: map[string]*structpb.Value{
-				model.TLSModeLabelShortname: {Kind: &structpb.Value_StringValue{StringValue: tlsMode}},
-			},
-		}
-	} else {
-		delete(newEndpoint.Metadata.FilterMetadata, EnvoyTransportSocketMetadataKey)
-	}
-	return newEndpoint, true
 }
 
 func addIstioEndpointLabel(metadata *core.Metadata, key string, val *structpb.Value) {