--- a/pilot/pkg/security/authn/policy_applier.go
+++ b/pilot/pkg/security/authn/policy_applier.go
@@ -18,7 +18,6 @@ import (
 	hcm "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/network/http_connection_manager/v3"
 	tlsv3 "github.com/envoyproxy/go-control-plane/envoy/extensions/transport_sockets/tls/v3"
 
-	"istio.io/api/security/v1beta1"
 	"istio.io/istio/pilot/pkg/model"
 )
 
@@ -40,7 +39,7 @@ type PolicyApplier interface {
 	AuthNFilter(forSidecar bool) *hcm.HttpFilter
 
 	// PortLevelSetting returns port level mTLS settings.
-	PortLevelSetting() map[uint32]v1beta1.PeerAuthentication_MutualTLS_Mode
+	PortLevelSetting() map[uint32]model.MutualTLSMode
 
 	MtlsPolicy
 }