--- a/pilot/pkg/xds/endpoint_builder.go
+++ b/pilot/pkg/xds/endpoint_builder.go
@@ -388,7 +388,7 @@ func buildEnvoyLbEndpoint(b *EndpointBuilder, e *model.IstioEndpoint) *endpoint.
 
 	// detect if mTLS is possible for this endpoint, used later during ep filtering
 	// this must be done while converting IstioEndpoints because we still have workload labels
-	if !b.mtlsChecker.computeMtlsEnabled(e) {
+	if !b.mtlsChecker.checkMtlsEnabled(e) {
 		meta.TLSMode = ""
 	}
 	util.AppendLbEndpointMetadata(meta, ep.Metadata)