--- a/pilot/pkg/xds/endpoint_builder.go
+++ b/pilot/pkg/xds/endpoint_builder.go
@@ -295,11 +295,21 @@ func (b *EndpointBuilder) buildLocalityLbEndpointsFromShards(
 				}
 			}
 
-			// Currently the HBONE implementation leads to different endpoint generation depending on if the
-			// client proxy supports HBONE or not. This breaks the cache.
-			// For now, just disable caching if the global HBONE flag is enabled.
-			if ep.EnvoyEndpoint == nil || features.EnableHBONE {
-				eep := buildEnvoyLbEndpoint(b, ep)
+			mtlsEnabled := b.mtlsChecker.checkMtlsEnabled(ep)
+			// Determine if we need to build the endpoint. We try to cache it for performance reasons
+			needToCompute := ep.EnvoyEndpoint == nil
+			if features.EnableHBONE {
+				// Currently the HBONE implementation leads to different endpoint generation depending on if the
+				// client proxy supports HBONE or not. This breaks the cache.
+				// For now, just disable caching if the global HBONE flag is enabled.
+				needToCompute = true
+			}
+			if ep.EnvoyEndpoint != nil && mtlsEnabled != isMtlsEnabled(ep.EnvoyEndpoint) {
+				// The mTLS settings may have changed, invalidating the cache endpoint. Rebuild it
+				needToCompute = true
+			}
+			if needToCompute {
+				eep := buildEnvoyLbEndpoint(b, ep, mtlsEnabled)
 				if eep == nil {
 					continue
 				}
@@ -361,7 +371,7 @@ func (b *EndpointBuilder) createClusterLoadAssignment(llbOpts []*LocalityEndpoin
 }
 
 // buildEnvoyLbEndpoint packs the endpoint based on istio info.
-func buildEnvoyLbEndpoint(b *EndpointBuilder, e *model.IstioEndpoint) *endpoint.LbEndpoint {
+func buildEnvoyLbEndpoint(b *EndpointBuilder, e *model.IstioEndpoint, mtlsEnabled bool) *endpoint.LbEndpoint {
 	addr := util.BuildAddress(e.Address, e.EndpointPort)
 	healthStatus := e.HealthStatus
 	if features.DrainingLabel != "" && e.Labels[features.DrainingLabel] != "" {
@@ -388,7 +398,7 @@ func buildEnvoyLbEndpoint(b *EndpointBuilder, e *model.IstioEndpoint) *endpoint.
 
 	// detect if mTLS is possible for this endpoint, used later during ep filtering
 	// this must be done while converting IstioEndpoints because we still have workload labels
-	if !b.mtlsChecker.checkMtlsEnabled(e) {
+	if !mtlsEnabled {
 		meta.TLSMode = ""
 	}
 	util.AppendLbEndpointMetadata(meta, ep.Metadata)