* Remove all caching. This is expensive and offers little benefit; the
  issue here is the mtlsChecker is built per endpointBuilder, which is
per-proxy-cluster. So it has very little re-use. We find cache key
computation is actually way more expensive than actually computing the
value anyways. Note - when it was first written this probably wasn't the case, we optimized it after
* Simplify DR processing. DR is a property of the cluster, so do not
  compute per endpoint
* Fix DR processing. We were using subset rule when not creating subset
  cluster. This changes the tests; everything else is a NOP.
* Micro-optimize a few hot paths in the code now that we are calling
  them a bit more.

end result (edit: see https://github.com/istio/istio/pull/45230#discussion_r1213281992 for updated numbers)
```
name                          old time/op        new time/op        delta
EndpointGeneration/1/100-48          634µs ± 1%         606µs ± 1%   -4.49%  (p=0.008 n=5+5)
EndpointGeneration/10/10-48          306µs ± 2%         285µs ± 1%   -6.89%  (p=0.008 n=5+5)
EndpointGeneration/100/10-48        2.71ms ± 1%        2.51ms ± 2%   -7.41%  (p=0.008 n=5+5)
EndpointGeneration/1000/1-48        2.69ms ± 2%        2.45ms ± 2%   -9.23%  (p=0.008 n=5+5)

name                          old alloc/op       new alloc/op       delta
EndpointGeneration/1/100-48          219kB ± 0%         186kB ± 0%  -14.97%  (p=0.008 n=5+5)
EndpointGeneration/10/10-48         79.9kB ± 0%        73.7kB ± 0%   -7.75%  (p=0.008 n=5+5)
EndpointGeneration/100/10-48         672kB ± 0%         617kB ± 0%   -8.29%  (p=0.008 n=5+5)
EndpointGeneration/1000/1-48         696kB ± 0%         610kB ± 0%  -12.34%  (p=0.008 n=5+5)

name                          old allocs/op      new allocs/op      delta
EndpointGeneration/1/100-48          4.41k ± 0%         4.01k ± 0%   -9.07%  (p=0.008 n=5+5)
EndpointGeneration/10/10-48          1.66k ± 0%         1.61k ± 0%   -3.02%  (p=0.008 n=5+5)
EndpointGeneration/100/10-48         12.7k ± 0%         12.5k ± 0%   -1.01%  (p=0.008 n=5+5)
EndpointGeneration/1000/1-48         12.2k ± 0%         12.1k ± 0%   -0.34%  (p=0.000 n=5+4)
```

One note is this under counts the benefit because the benchmark config is fairly simple. In a real world
environment the impact is much larger, see https://github.com/istio/istio/issues/44985