@istio-testing 2023-05-31
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
unit-tests-arm64_istio | 6980aad663974ee5de01caf1b3e35bade5f9f3fb | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/45230/unit-tests-arm64_istio/1664036066237091840) | true | `/test unit-tests-arm64`
gencheck_istio | 6980aad663974ee5de01caf1b3e35bade5f9f3fb | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/45230/gencheck_istio/1664036078555762688) | true | `/test gencheck`
lint_istio | 6980aad663974ee5de01caf1b3e35bade5f9f3fb | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/45230/lint_istio/1664036077708513280) | true | `/test lint`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->